# swagger-spec.json
{"openapi":"3.0.0","paths":{"/health":{"get":{"operationId":"AppController_getHealth","parameters":[],"responses":{"200":{"description":""}}}},"/user/signup":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["user"]}},"/user/login":{"post":{"operationId":"UserController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"201":{"description":""}},"tags":["user"]}},"/user/confirm":{"get":{"operationId":"UserController_confirm","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/site":{"post":{"operationId":"SiteController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSiteDto"}}}},"responses":{"201":{"description":""}},"tags":["site"],"security":[{"bearer":[]}]},"get":{"operationId":"SiteController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["site"],"security":[{"bearer":[]}]}},"/site/search":{"post":{"operationId":"SiteController_findReportByTags","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchSiteReportDto"}}}},"responses":{"201":{"description":""}},"tags":["site"],"security":[{"bearer":[]}]}},"/site/reports":{"get":{"operationId":"SiteController_getReports","parameters":[],"responses":{"200":{"description":""}},"tags":["site"],"security":[{"bearer":[]}]}},"/site/{id}/report":{"get":{"operationId":"SiteController_getReport","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["site"],"security":[{"bearer":[]}]}},"/site/{id}":{"get":{"operationId":"SiteController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["site"],"security":[{"bearer":[]}]},"patch":{"operationId":"SiteController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSiteDto"}}}},"responses":{"200":{"description":""}},"tags":["site"],"security":[{"bearer":[]}]},"delete":{"operationId":"SiteController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["site"],"security":[{"bearer":[]}]}},"/site-report":{"get":{"operationId":"SiteReportController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["site-report"],"security":[{"bearer":[]}]}},"/site-report/{id}":{"get":{"operationId":"SiteReportController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["site-report"],"security":[{"bearer":[]}]}}},"info":{"title":"WATCH-SITE Server API","description":"RESTFUL Backend server","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["name","email","password"]},"LoginUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateSiteDto":{"type":"object","properties":{"name":{"type":"string","example":"My website","description":"The name of the check."},"url":{"type":"string","example":"https://www.google.com","description":"The URL to be monitored."},"protocol":{"type":"string","enum":["http","https","tcp"],"example":"http","description":"The resource protocol name."},"path":{"type":"string","example":"/health-check","description":"A specific path to be monitored."},"port":{"type":"number","example":80,"description":"The server port number."},"webhook":{"type":"string","example":"https://mywebhook.com","description":"A webhook URL to receive a notification on."},"timeout":{"type":"number","example":5000,"description":"The timeout of the polling request."},"interval":{"type":"number","example":5,"description":"The time interval (minutes) for polling requests. (min: 5 mins)"},"threshold":{"type":"number","example":3,"description":"The threshold of failed requests that will create an alert."},"authUsername":{"type":"string","description":"The authentication username."},"authPassword":{"type":"string","description":"The authentication password."},"httpHeaders":{"example":[["Content-Type","application/json"],["Authorization","Bearer token"]],"type":"array","items":{"type":"string"}},"assertCode":{"type":"number","description":"The response assertion to be used on the polling response."},"tags":{"example":["tag1","tag2"],"description":"A list of the check tags.","type":"array","items":{"type":"string"}},"ignoreSSL":{"type":"boolean","example":true,"description":"A flag to ignore broken/expired SSL certificates in case of using the HTTPS protocol."}},"required":["name","url","protocol","path","port","webhook","timeout","interval","threshold","authUsername","authPassword","httpHeaders","assertCode","tags","ignoreSSL"]},"SearchSiteReportDto":{"type":"object","properties":{"tags":{"description":"Tags to search for","type":"array","items":{"type":"string"}}},"required":["tags"]},"UpdateSiteDto":{"type":"object","properties":{"name":{"type":"string","example":"My website","description":"The name of the check."},"url":{"type":"string","example":"https://www.google.com","description":"The URL to be monitored."},"protocol":{"type":"string","enum":["http","https","tcp"],"example":"http","description":"The resource protocol name."},"path":{"type":"string","example":"/health-check","description":"A specific path to be monitored."},"port":{"type":"number","example":80,"description":"The server port number."},"webhook":{"type":"string","example":"https://mywebhook.com","description":"A webhook URL to receive a notification on."},"timeout":{"type":"number","example":5000,"description":"The timeout of the polling request."},"interval":{"type":"number","example":5,"description":"The time interval (minutes) for polling requests. (min: 5 mins)"},"threshold":{"type":"number","example":3,"description":"The threshold of failed requests that will create an alert."},"authUsername":{"type":"string","description":"The authentication username."},"authPassword":{"type":"string","description":"The authentication password."},"httpHeaders":{"example":[["Content-Type","application/json"],["Authorization","Bearer token"]],"type":"array","items":{"type":"string"}},"assertCode":{"type":"number","description":"The response assertion to be used on the polling response."},"tags":{"example":["tag1","tag2"],"description":"A list of the check tags.","type":"array","items":{"type":"string"}},"ignoreSSL":{"type":"boolean","example":true,"description":"A flag to ignore broken/expired SSL certificates in case of using the HTTPS protocol."}}}}}}